refactoring: Removing KafkaDataSigner and KafkaDataVerifier and relying on new GenericSignerVerifier class and Broker interface to simplify creation of Signer (formerly DataSigner) and Verifier (formerly DataVerifier) implementations.
