+++ java/src/main/java/io/github/cyfko/dverify/TokenMode.java
@@ -1,0 +1,6 @@
+package io.github.cyfko.dverify;
+
+public enum TokenMode {
+    jwt,
+    uuid,
+}
