+++ java/src/test/java/io/github/cyfko/dverify/VerifierIdentityTest.java
@@ -1,12 +1,10 @@
 package io.github.cyfko.dverify;
 
-
 import io.github.cyfko.dverify.exceptions.DataExtractionException;
 import io.github.cyfko.dverify.exceptions.JsonEncodingException;
-import io.github.cyfko.dverify.impl.kafka.KafkaDataSigner;
-import io.github.cyfko.dverify.impl.kafka.KafkaDataVerifier;
+import io.github.cyfko.dverify.impl.GenericSignerVerifier;
+import io.github.cyfko.dverify.impl.kafka.*;
 
-import io.github.cyfko.dverify.impl.kafka.VerifierConfig;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.junit.jupiter.api.*;
 import org.testcontainers.containers.KafkaContainer;
@@ -17,15 +15,14 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.time.Duration;
 import java.util.Properties;
-import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class DataVerifierJwtTest {
+public class VerifierIdentityTest {
 
     private static KafkaContainer kafkaContainer;
-    private DataSigner signer;
-    private DataVerifier verifier;
+    private Signer signer;
+    private Verifier verifier;
     private File tempDir;
 
     @BeforeAll
@@ -45,13 +42,14 @@ public class DataVerifierJwtTest {
         String kafkaBootstrapServers = kafkaContainer.getBootstrapServers();
 
         Properties props = new Properties();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrapServers);
+        props.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrapServers);
 
         tempDir = Files.createTempDirectory("rocksdb_db_test_").toFile();
-        props.put(VerifierConfig.EMBEDDED_DB_PATH_CONFIG, tempDir.getAbsolutePath());
+        props.setProperty(VerifierConfig.EMBEDDED_DB_PATH_CONFIG, tempDir.getAbsolutePath());
 
-        signer = KafkaDataSigner.of(props);
-        verifier = KafkaDataVerifier.of(props);
+        GenericSignerVerifier genericSignerVerifier = new GenericSignerVerifier(KafkaBrokerAdapter.of(props));
+        signer = genericSignerVerifier;
+        verifier = genericSignerVerifier;
     }
 
     @AfterEach
@@ -66,7 +64,7 @@ public class DataVerifierJwtTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofHours(2);
 
-        String jwt = signer.sign(data, duration);
+        String jwt = signer.sign(data, duration, TokenMode.jwt);
 
         assertNotNull(jwt);
         assertFalse(jwt.isEmpty());
@@ -77,7 +75,7 @@ public class DataVerifierJwtTest {
         Object invalidData = null; // Simulating invalid data
         Duration duration = Duration.ofHours(2);
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, TokenMode.jwt));
     }
 
     @Test
@@ -85,7 +83,7 @@ public class DataVerifierJwtTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofMinutes(-5); // Negative duration
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, duration));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, duration, TokenMode.jwt));
     }
 
     @Test
@@ -93,7 +91,7 @@ public class DataVerifierJwtTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofHours(2);
 
-        String jwt = signer.sign(data, duration);
+        String jwt = signer.sign(data, duration, TokenMode.jwt);
 
         assertNotNull(jwt, "JWT should not be null");
         assertFalse(jwt.isEmpty(), "JWT should not be empty");
@@ -104,14 +102,14 @@ public class DataVerifierJwtTest {
         Object invalidData = null; // Simulating invalid data
         Duration duration = Duration.ofHours(2);
 
-        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration));
+        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, TokenMode.jwt));
     }
 
     @Test
     public void verify_valid_token_should_returns_payload() throws InterruptedException {
         UserData data = new UserData("john.doe@example.com");
-        String jwt = signer.sign(data, Duration.ofHours(2)); // Generate a valid token
-        Thread.sleep(10); // Wait 10 ms to ensure that the keys has been propagated to kafka
+        String jwt = signer.sign(data, Duration.ofHours(2), TokenMode.jwt); // Generate a valid token
+        Thread.sleep(5000); // Wait 5 seconds to ensure that the keys has been propagated to kafka
 
         UserData verifiedData = verifier.verify(jwt, UserData.class);
 
@@ -129,7 +127,7 @@ public class DataVerifierJwtTest {
     @Test
     public void verify_expired_token_should_throws_exception() throws InterruptedException {
         UserData data = new UserData("john.doe@example.com");
-        String jwt = signer.sign(data, Duration.ofMillis(1)); // Token with short duration
+        String jwt = signer.sign(data, Duration.ofMillis(1), TokenMode.jwt); // Token with short duration
         Thread.sleep(10); // Wait for the token to expire
 
         assertThrows(DataExtractionException.class, () -> verifier.verify(jwt, UserData.class));
