+++ java/src/test/java/io/github/cyfko/dverify/VerifierJwtTest.java
@@ -1,9 +1,12 @@
 package io.github.cyfko.dverify;
 
+
 import io.github.cyfko.dverify.exceptions.DataExtractionException;
 import io.github.cyfko.dverify.exceptions.JsonEncodingException;
-import io.github.cyfko.dverify.impl.kafka.*;
+import io.github.cyfko.dverify.impl.GenericSignerVerifier;
+import io.github.cyfko.dverify.impl.kafka.KafkaBrokerAdapter;
 
+import io.github.cyfko.dverify.impl.kafka.VerifierConfig;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.junit.jupiter.api.*;
 import org.testcontainers.containers.KafkaContainer;
@@ -17,11 +20,11 @@ import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class DataVerifierIdentityTest {
+public class VerifierJwtTest {
 
     private static KafkaContainer kafkaContainer;
-    private DataSigner signer;
-    private DataVerifier verifier;
+    private Signer signer;
+    private Verifier verifier;
     private File tempDir;
 
     @BeforeAll
@@ -41,15 +44,14 @@ public class DataVerifierIdentityTest {
         String kafkaBootstrapServers = kafkaContainer.getBootstrapServers();
 
         Properties props = new Properties();
-        props.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrapServers);
-        props.setProperty(SignerConfig.GENERATED_TOKEN_CONFIG, Constant.GENERATED_TOKEN_IDENTITY);
+        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrapServers);
 
         tempDir = Files.createTempDirectory("rocksdb_db_test_").toFile();
-        props.setProperty(VerifierConfig.EMBEDDED_DB_PATH_CONFIG, tempDir.getAbsolutePath());
-        props.setProperty(VerifierConfig.CLEANUP_INTERVAL_CONFIG, "1");
+        props.put(VerifierConfig.EMBEDDED_DB_PATH_CONFIG, tempDir.getAbsolutePath());
 
-        signer = KafkaDataSigner.of(props);
-        verifier = KafkaDataVerifier.of(props);
+        GenericSignerVerifier genericSignerVerifier = new GenericSignerVerifier(KafkaBrokerAdapter.of(props));
+        signer = genericSignerVerifier;
+        verifier = genericSignerVerifier;
     }
 
     @AfterEach
@@ -64,7 +66,7 @@ public class DataVerifierIdentityTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofHours(2);
 
-        String jwt = signer.sign(data, duration);
+        String jwt = signer.sign(data, duration, TokenMode.jwt);
 
         assertNotNull(jwt);
         assertFalse(jwt.isEmpty());
@@ -75,7 +77,7 @@ public class DataVerifierIdentityTest {
         Object invalidData = null; // Simulating invalid data
         Duration duration = Duration.ofHours(2);
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, TokenMode.jwt));
     }
 
     @Test
@@ -83,7 +85,7 @@ public class DataVerifierIdentityTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofMinutes(-5); // Negative duration
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, duration));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, duration, TokenMode.jwt));
     }
 
     @Test
@@ -91,7 +93,7 @@ public class DataVerifierIdentityTest {
         UserData data = new UserData("john.doe@example.com");
         Duration duration = Duration.ofHours(2);
 
-        String jwt = signer.sign(data, duration);
+        String jwt = signer.sign(data, duration, TokenMode.jwt);
 
         assertNotNull(jwt, "JWT should not be null");
         assertFalse(jwt.isEmpty(), "JWT should not be empty");
@@ -102,14 +104,14 @@ public class DataVerifierIdentityTest {
         Object invalidData = null; // Simulating invalid data
         Duration duration = Duration.ofHours(2);
 
-        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration));
+        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, TokenMode.jwt));
     }
 
     @Test
     public void verify_valid_token_should_returns_payload() throws InterruptedException {
         UserData data = new UserData("john.doe@example.com");
-        String jwt = signer.sign(data, Duration.ofHours(2)); // Generate a valid token
-        Thread.sleep(10); // Wait 10 ms to ensure that the keys has been propagated to kafka
+        String jwt = signer.sign(data, Duration.ofHours(2), TokenMode.jwt); // Generate a valid token
+        Thread.sleep(5000); // Wait 5 secs to ensure that the keys has been propagated to kafka
 
         UserData verifiedData = verifier.verify(jwt, UserData.class);
 
@@ -127,7 +129,7 @@ public class DataVerifierIdentityTest {
     @Test
     public void verify_expired_token_should_throws_exception() throws InterruptedException {
         UserData data = new UserData("john.doe@example.com");
-        String jwt = signer.sign(data, Duration.ofMillis(1)); // Token with short duration
+        String jwt = signer.sign(data, Duration.ofMillis(1), TokenMode.jwt); // Token with short duration
         Thread.sleep(10); // Wait for the token to expire
 
         assertThrows(DataExtractionException.class, () -> verifier.verify(jwt, UserData.class));
